+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "homebridge-august-smart-locks",
-  "version": "0.4.5",
+  "version": "0.4.6",
   "description": "August Lock plugin for HomeBridge: https://github.com/nfarina/homebridge",
   "repository": {
     "type": "git",
