+++ settings.js
@@ -1,0 +1,10 @@
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.PLUGIN_NAME = exports.PLATFORM_NAME = void 0;
+/**
+ * This is the name of the platform that users will use to register the plugin in the Homebridge config.json
+ */
+exports.PLATFORM_NAME = 'AugustLocks';
+/**
+ * This must match the name of your plugin as defined the package.json
+ */
+exports.PLUGIN_NAME = 'homebridge-august-smart-locks';
